refactor(workflow): 合并重复的状态面板组件
解决workflow右侧顶部重复组件的问题：

- 新增UnifiedStatusPanel组件，合并ExecutionControlPanel和WorkflowStatus功能

- 统一显示执行状态、进度条、操作按钮和执行时间

- 删除重复的ExecutionControlPanel和WorkflowStatus组件

- 优化布局层次，减少视觉冗余

- 保持所有原有功能：停止、重试、重置、查看结果

- 改进状态信息展示，错误信息更清晰

- 统一的stone色系设计，保持视觉一致性
